// edci-viewer/edci-viewer-web/src/main/angular/src/app/shared/swagger/model/europassCredentialPresentationView.ts
/**
 * API
 * API Swagger description
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AchievementTabView } from './achievementTabView';
import { ActivityTabView } from './activityTabView';
import { AssessmentTabView } from './assessmentTabView';
import { CredentialMetadataTabView } from './credentialMetadataTabView';
import { CredentialSubjectTabView } from './credentialSubjectTabView';
import { EntitlementTabView } from './entitlementTabView';
import { EuropassCredentialPresentationLiteView } from './europassCredentialPresentationLiteView';
import { OrganizationTabView } from './organizationTabView';


export interface EuropassCredentialPresentationView { 
    credentialMetadata?: CredentialMetadataTabView;
    credentialSubject?: CredentialSubjectTabView;
    issuerPresentation?: OrganizationTabView;
    issuerCredential?: OrganizationTabView;
    subCredentials?: Array<EuropassCredentialPresentationLiteView>;
    achievements?: Array<AchievementTabView>;
    activities?: Array<ActivityTabView>;
    entitlements?: Array<EntitlementTabView>;
    assessments?: Array<AssessmentTabView>;
    achievementsList?: Array<AchievementTabView>;
    activitiesList?: Array<ActivityTabView>;
    entitlementsList?: Array<EntitlementTabView>;
    organizationsList?: Array<OrganizationTabView>;
    assessmentsList?: Array<AssessmentTabView>;
}