// edci-viewer/edci-viewer-web/src/main/angular/src/app/shared/swagger/model/credentialSubjectTabView.ts
/**
 * API
 * API Swagger description
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { ContactPointFieldView } from './contactPointFieldView';
import { GroupFieldView } from './groupFieldView';
import { IdentifierFieldView } from './identifierFieldView';
import { LocationFieldView } from './locationFieldView';
import { NoteFieldView } from './noteFieldView';
import { OrganizationTabView } from './organizationTabView';


export interface CredentialSubjectTabView { 
    altLabel?: Array<string>;
    contactPoint?: Array<ContactPointFieldView>;
    groupMemberOf?: Array<GroupFieldView>;
    identifier?: Array<IdentifierFieldView>;
    dateModified?: string;
    locations?: Array<LocationFieldView>;
    additionalNote?: Array<NoteFieldView>;
    prefLabel?: Array<string>;
    legalName?: string;
    eidasIdentifier?: IdentifierFieldView;
    vatIdentifier?: Array<IdentifierFieldView>;
    taxIdentifier?: Array<IdentifierFieldView>;
    nationalID?: IdentifierFieldView;
    fullName?: string;
    registration?: IdentifierFieldView;
    id?: string;
    givenName?: string;
    familyName?: string;
    patronymicName?: string;
    birthName?: string;
    dateOfBirth?: string;
    citizenshipCountry?: Array<string>;
    placeOfBirth?: LocationFieldView;
    gender?: string;
    memberOf?: Array<OrganizationTabView>;
}